se retorna page.js a su estado anterior al ejercicio de prueba DOM
